+++ tomcat/src/main/java/org/apache/coyote/http11/handler/exception/HttpMethodInvalidException.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.exception;
+package org.apache.coyote.http11.handler.exception;
 
 public class HttpMethodInvalidException extends RuntimeException {
