+++ tomcat/src/main/java/org/apache/coyote/http11/request/uri/HttpMethod.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.request.uri;
 
-import org.apache.coyote.http11.exception.HttpMethodInvalidException;
+import org.apache.coyote.http11.handler.exception.HttpMethodInvalidException;
 
 import java.util.Arrays;
